+++ hardware/P1MeterTCP.cpp
@@ -169,7 +169,7 @@ void P1MeterTCP::OnDisconnect()
 
 void P1MeterTCP::OnData(const unsigned char *pData, size_t length)
 {
-	ParseP1Data((const unsigned char*)pData, length, m_bDisableCRC, m_ratelimit);
+	ParseP1Data((const unsigned char*)pData, static_cast<int>(length), m_bDisableCRC, m_ratelimit);
 }
