+++ hardware/SatelIntegra.cpp
@@ -1116,7 +1116,7 @@ int SatelIntegra::SendCommand(const unsigned char* cmd, const unsigned int cmdLe
 	struct timeval tv;
 	tv.tv_sec = 3;
 	tv.tv_usec = 0;
-	if (select(m_socket + 1, &rfds, nullptr, nullptr, &tv) < 0)
+	if (select((int)m_socket + 1, &rfds, nullptr, nullptr, &tv) < 0)
 	{
 		Log(LOG_ERROR, "connection lost.");
 		DestroySocket();
@@ -1228,7 +1228,7 @@ std::pair<unsigned char*, unsigned int> SatelIntegra::getFullFrame(const unsigne
 	result.push_back(0xFE);
 	result.push_back(0x0D);
 
-	unsigned int resultSize = result.size();
+	size_t resultSize = result.size();
 	unsigned char* pResult = new unsigned char[resultSize];
 	memset(pResult, 0, resultSize);
 	std::list<unsigned char>::iterator it = result.begin();
@@ -1237,5 +1237,5 @@ std::pair<unsigned char*, unsigned int> SatelIntegra::getFullFrame(const unsigne
 		pResult[index] = *it;
 	}
 
-	return { pResult, resultSize };
+	return { pResult, static_cast<unsigned int>(resultSize) };
 }
